Changed vec2 class to work with ints

+++ intelligame/intelligame/src/main/kotlin/com/glycin/intelligame/pong/model/PlayerBrick.kt
@@ -19,11 +19,11 @@ class PlayerBrick(
     private val speed: Float = 2f
 
     fun moveUp(deltaTime: Float) {
-        position += Vec2.up * (deltaTime * speed)
+        position += Vec2.up * (deltaTime * speed).toInt()
     }
 
     fun moveDown(deltaTime: Float) {
-        position += Vec2.down * (deltaTime * speed)
+        position += Vec2.down * (deltaTime * speed).toInt()
     }
 
     override fun getCollisionNormal(colPosition: Vec2): Vec2 {
