+++ intelligame/intelligame/src/main/kotlin/com/glycin/intelligame/shared/Vec2.kt
@@ -3,30 +3,30 @@ package com.glycin.intelligame.shared
 import kotlin.math.sqrt
 
 data class Vec2(
-    val x: Float = 0f,
-    val y: Float = 0f,
+    val x: Int = 0,
+    val y: Int = 0,
 ) {
     
     companion object {
-        val zero = Vec2(0f, 0f)
-        val one = Vec2(1f, 1f)
-        val up = Vec2(0f, -1f)
-        val down = Vec2(0f, 1f)
-        val left = Vec2(-1f, 0f)
-        val right = Vec2(1f, 0f)
+        val zero = Vec2(0, 0)
+        val one = Vec2(1, 1)
+        val up = Vec2(0, -1)
+        val down = Vec2(0, 1)
+        val left = Vec2(-1, 0)
+        val right = Vec2(1, 0)
     }
 
     operator fun plus(other: Vec2) = Vec2(x + other.x, y + other.y)
     operator fun minus(other: Vec2) = Vec2(x - other.x, y - other.y)
-    operator fun times(scalar: Float) = Vec2(x * scalar, y * scalar)
-    operator fun div(scalar: Float) = Vec2(x / scalar, y / scalar)
+    operator fun times(scalar: Int) = Vec2(x * scalar, y * scalar)
+    operator fun div(scalar: Int) = Vec2(x / scalar, y / scalar)
 
     fun dot(other: Vec2) = x * other.x + y * other.y
 
-    fun magnitude() = sqrt(x * x + y * y)
+    fun magnitude() = sqrt(x.toDouble() * x + y * y).toInt()
 
     fun normalized(): Vec2 {
         val mag = magnitude()
-        return if (mag != 0f) this / mag else zero
+        return if (mag != 0) this / mag else zero
     }
 }
